mz-repr: move timezone code to mz-pgtz
A future commit will load timezone abbreviations into `enum Timezone`,
to support abbreviations that refer to IANA timezone names, rather than
fixed offsets from UTC.
