+++ src/expr/Cargo.toml
@@ -31,6 +31,7 @@ mz-lowertest = { path = "../lowertest" }
 mz-ore = { path = "../ore", features = ["network"] }
 mz-persist-types = { path = "../persist-types" }
 mz-pgrepr = { path = "../pgrepr" }
+mz-pgtz = { path = "../pgtz" }
 mz-proto = { path = "../proto" }
 mz-regexp = { path = "../regexp" }
 mz-repr = { path = "../repr" }
