+++ src/expr/src/scalar/func/impls/time.rs
@@ -11,7 +11,8 @@ use std::fmt;
 
 use chrono::{NaiveDateTime, NaiveTime, Offset, TimeZone, Timelike};
 use mz_lowertest::MzReflect;
-use mz_repr::adt::datetime::{DateTimeField, DateTimeUnits, Timezone};
+use mz_pgtz::timezone::Timezone;
+use mz_repr::adt::datetime::{DateTimeField, DateTimeUnits};
 use mz_repr::adt::interval::Interval;
 use mz_repr::adt::numeric::{DecimalLike, Numeric};
 use mz_repr::adt::timestamp::TimeLike;
