+++ src/pgtz/src/lib.rs
@@ -76,3 +76,4 @@
 // END LINT CONFIG
 
 pub mod abbrev;
+pub mod timezone;
