+++ src/pgtz/src/timezone.proto
@@ -1,0 +1,21 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+syntax = "proto3";
+
+package mz_pgtz.timezone;
+
+import "proto/src/chrono.proto";
+
+message ProtoTimezone {
+    oneof kind {
+        mz_proto.chrono.ProtoFixedOffset fixed_offset = 1;
+        mz_proto.chrono.ProtoTz tz = 2;
+    }
+}
