+++ src/repr/Cargo.toml
@@ -33,6 +33,7 @@ once_cell = "1.16.0"
 mz-lowertest = { path = "../lowertest" }
 mz-ore = { path = "../ore", features = ["bytes", "smallvec", "stack", "test", "serde"] }
 mz-persist-types = { path = "../persist-types" }
+mz-pgtz = { path = "../pgtz" }
 mz-proto = { path = "../proto", features = ["chrono"] }
 num-traits = "0.2.15"
 num_enum = "0.5.7"
