+++ src/repr/src/adt/datetime.proto
@@ -11,15 +11,7 @@ syntax = "proto3";
 
 package mz_repr.adt.datetime;
 
-import "proto/src/chrono.proto";
 import "google/protobuf/empty.proto";
-
-message ProtoTimezone {
-    oneof kind {
-        mz_proto.chrono.ProtoFixedOffset fixed_offset = 1;
-        mz_proto.chrono.ProtoTz tz = 2;
-    }
-}
 
 message ProtoDateTimeUnits {
     oneof kind {
