+++ src/repr/src/strconv.rs
@@ -39,6 +39,7 @@ use mz_ore::error::ErrorExt;
 use mz_ore::fmt::FormatBuffer;
 use mz_ore::lex::LexBuf;
 use mz_ore::str::StrExt;
+use mz_pgtz::timezone::{Timezone, TimezoneSpec};
 use mz_proto::{RustType, TryFromProtoError};
 use num_traits::Float as NumFloat;
 use once_cell::sync::Lazy;
@@ -50,7 +51,7 @@ use uuid::Uuid;
 
 use crate::adt::array::ArrayDimension;
 use crate::adt::date::Date;
-use crate::adt::datetime::{self, DateTimeField, ParsedDateTime, Timezone, TimezoneSpec};
+use crate::adt::datetime::{self, DateTimeField, ParsedDateTime};
 use crate::adt::interval::Interval;
 use crate::adt::jsonb::{Jsonb, JsonbRef};
 use crate::adt::mz_acl_item::{AclItem, MzAclItem};
@@ -367,7 +368,7 @@ where
 /// <time zone interval> ::=
 ///     <sign> <hours value> <colon> <minutes value>
 /// ```
-fn parse_timestamp_string(s: &str) -> Result<(NaiveDate, NaiveTime, datetime::Timezone), String> {
+fn parse_timestamp_string(s: &str) -> Result<(NaiveDate, NaiveTime, Timezone), String> {
     if s.is_empty() {
         return Err("timestamp string is empty".into());
     }
@@ -393,7 +394,7 @@ fn parse_timestamp_string(s: &str) -> Result<(NaiveDate, NaiveTime, datetime::Ti
     let offset = if tz_string.is_empty() {
         Default::default()
     } else {
-        Timezone::parse(tz_string, TimezoneSpec::ISO)?
+        Timezone::parse(tz_string, TimezoneSpec::Iso)?
     };
 
     Ok((d, t, offset))
@@ -472,7 +473,7 @@ where
 pub fn parse_timestamptz(s: &str) -> Result<CheckedTimestamp<DateTime<Utc>>, ParseError> {
     parse_timestamp_string(s)
         .and_then(|(date, time, timezone)| {
-            use datetime::Timezone::*;
+            use Timezone::*;
             let mut dt = date.and_time(time);
             let offset = match timezone {
                 FixedOffset(offset) => offset,
